+++ ChangeLog.md
@@ -1,5 +1,16 @@
 # Braindrop ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Updated the help screen so that the command table shows the command name
+  first. ([#53](https://github.com/davep/braindrop/pull/53))
+- Updated the help screen so that the commands are sorted by the command
+  name. ([#53](https://github.com/davep/braindrop/pull/53))
+- Various cosmetic tweaks to the help screen.
+  ([#53](https://github.com/davep/braindrop/pull/53))
+
 ## v0.3.0
 
 **Released: 2025-01-08**